Added API and plugin structure
As bronx will have a public API, this commit adds the basic API and
plugin architecture. The two maven modules are no linked to its parent
as they technically need to be available as standalone projects.

This commit also publishes the bronx plugin instance to the spigot
service manager when the plugin is loaded.